Lazy load auto-updater
